Remove chmod for kdb/db files being copied on firstboot.
Due to changes in default file permissions and libsqlfs being able to
handle cp -a correctly, we no longer need to recursively chmod the
sqlfs settings db. Regardless, busybox chmod doesn't support the option
we were using and we needed to change due to that anyway.

Also fix the libsqlfs.service to require the setup package and update
the setup.service to no longer fail if modprobe fuse doesn't work.

Signed-off-by: William Douglas <william.douglas@intel.com>