Fix directory permissions of sqlfs mount point after mounting
Without doing a chmod here, the mounted sqlfs has permissions 0700,
which is too restrictive (user applications cannot read or write keys to
the sqlfs).

Modify the permissions to 1777, equivalent to permissions set in the sqlfs
initscript with the addition of the sticky bit.

Now applications are able to read/write to the sqlfs as before.

Change-Id: Iff23e4102fea682997c83e0ea3fc127d1cc6b030
Signed-off-by: Patrick McCarty <patrick.mccarty@linux.intel.com>